spendsense-ai-backend/backend/requirements.txt
# --- API & BACKEND ---
fastapi
uvicorn
pydantic

# --- FRONTEND ---
streamlit

# --- AI & LANGCHAIN ---
langchain
langchain-community
langchain-google-genai

# --- AGENT TOOLS ---
tavily-python

# --- DATABASE & DATA HANDLING ---
faiss-cpu
pandas

# --- UTILITIES ---
python-dotenv

ipykernel